Split native app container from core implementation
